+++ lib/xing_api/error.rb
@@ -1,16 +1,17 @@
 module XingApi
   class Error < StandardError
-    attr_reader :status_code, :name, :text
+    attr_reader :status_code, :name, :text, :errors
 
-    def initialize(status_code, name = '', text = '')
+    def initialize(status_code, name = '', text = '', errors = [])
       @status_code = status_code
       @name = name
       @text = text
+      @errors = errors || []
       super(text)
     end
 
     def to_s
-      [status_code, name, text].join(' - ')
+      [status_code, name, text, errors.to_s].join(' - ')
     end
   end